plugin levenberg marquardt test is split up to two tests
