+++ test/plugin_levenberg_marquardt/PluginLevenbergMarquardtTests.cpp
@@ -1,0 +1,5 @@
+#include "PluginLevenbergMarquardtTests.h"
+
+PluginLevenbergMarquardtTests::PluginLevenbergMarquardtTests() {
+    pluginsModelsDir = rrTestModelsDir_ / "PLUGINS";
+}
